+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2014-06-10
+
+Bugfixes:
+
+  - Using Guid.Empty instead of TaskInfo.QDSID to prevent "Could not find a result for the specified execution ID"
+
 ## 2013-10-11
 
 Bugfixes: